# backend/gateway/openapi.yaml
swagger: '2.0'
info:
  title: Semper API
  description: >
    API Gateway front door for the private Semper Cloud Run service.
    Device-attested routes require headers X-Device-Id, X-Nonce, X-Signature
    (ECDSA-P256 over nonce+METHOD+path+sha256(body)) in addition to the
    end-user Bearer ID token (Authorization or X-Forwarded-Authorization).

    POST /v1/tasks/provision-session is deliberately absent: it is the Cloud
    Tasks callback, delivered straight to the Cloud Run URL with an OIDC token
    for the task invoker service account, and must not be reachable through the
    public front door at all.
  version: '1.0'
schemes:
  - https
produces:
  - application/json
# All requests are forwarded to the private Cloud Run service, invoked with the
# gateway's backend-auth service account. The incoming path is appended to the
# address, so the backend sees the same path.
#
# Template — replace before `gcloud api-gateway api-configs create`:
#   CLOUD_RUN_URL=$(gcloud run services describe indic-api --region REGION \
#                     --format='value(status.url)')
#   # FIREBASE_PROJECT_ID must be the Auth project (e.g. indicvision-dic-app-auth),
#   # which can differ from GOOGLE_CLOUD_PROJECT (e.g. indicvision-dic-app).
#   sed -e "s|__CLOUD_RUN_URL__|${CLOUD_RUN_URL}|g" \
#       -e "s|__FIREBASE_PROJECT_ID__|${FIREBASE_PROJECT_ID}|g" \
#       openapi.yaml > openapi.generated.yaml
# Never commit a live Cloud Run hostname or project id here.
...
#
...
securityDefinitions:
  firebase:
    authorizationUrl: ''
    flow: implicit
    type: oauth2
    x-google-issuer: https://securetoken.google.com/__FIREBASE_PROJECT_ID__
    x-google-jwks_uri: https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com
    x-google-audiences: __FIREBASE_PROJECT_ID__
# Applies to every operation; the health probes opt out with `security: []`.
security:
  - firebase: []
# Per-consumer quotas (ESPv2 / Service Management). Deploy with the gateway
# OpenAPI config so these metrics are registered; in-process token buckets in
# app/rate_limit.py are belt-and-braces per Cloud Run instance.
...
paths:
...
  /v1/me:
    get:
      operationId: getMe
      responses:
        '200': { description: OK }
    delete:
      operationId: deleteAccount
      description: Device-attested GDPR erasure.
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
...
  /v1/me/export:
    get:
      operationId: exportAccount
      description: Device-attested GDPR data portability export.
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
...
  /v1/files/{fileId}/content:
    get:
      operationId: downloadFile
      description: >
        Device-attested download (same verified_device headers as writes).
        Signature covers GET + path + empty body; Range is an unsigned header.
        Session.zip is proxied through Cloud Run, so this route needs a longer
        deadline than the top-level 60s JSON budget.
      produces:
        - application/octet-stream
      # Operation-level backend MUST restate address + APPEND_PATH_TO_ADDRESS —
      # omitting path_translation defaults to CONSTANT_ADDRESS and breaks routing.
      x-google-backend:
        address: __CLOUD_RUN_URL__
        jwt_audience: __CLOUD_RUN_URL__
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 300.0
      x-google-quota:
        metricCosts:
          download-requests: 1
      parameters:
        - name: fileId
          in: path
          required: true
          type: string
        - name: Range
          in: header
          required: false
          type: string
          description: Optional byte range (e.g. bytes=1024-); proxied to Drive as 206.
      responses:
        '200': { description: OK }
        '206': { description: Partial content }
        '401': { description: Missing or invalid device attestation }
        '403': { description: feature_not_licensed (demo mode) }
        '416': { description: range_not_satisfiable }
        '429': { description: Rate limited }
        '502': { description: Drive download failed }
  /v1/admin/users:
    get:
      operationId: adminListUsers
      parameters:
        - name: status
          in: query
          required: false
          type: string
        - name: limit
          in: query
          required: false
          type: integer
          description: 1..1000 (default 200); clamped server-side.
      responses:
        '200': { description: OK }
  /v1/admin/users/{uid}/approve:
    post:
      operationId: adminApproveUser
      description: Device-attested admin mutation.
      parameters:
        - name: uid
          in: path
          required: true
          type: string
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
  /v1/admin/users/{uid}/revoke:
    post:
      operationId: adminRevokeUser
      description: Device-attested admin mutation.
      parameters:
        - name: uid
          in: path
          required: true
          type: string
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
  /v1/admin/users/{uid}/config:
    patch:
      operationId: adminPatchUserConfig
      description: >
        Device-attested. Set or clear per-user product-limit overrides
        (maxSessions, maxFilesPerSession, maxFrames).
      parameters:
        - name: uid
          in: path
          required: true
          type: string
      responses:
        '200': { description: OK }
        '400': { description: Empty patch }
        '401': { description: Missing or invalid device attestation }
        '404': { description: User not found }
...
  /v1/admin/licenses:
    get:
      operationId: adminListLicenses
      responses:
        '200': { description: OK }
    post:
      operationId: adminCreateLicense
      description: >
        Device-attested, Semper-staff only. Mint a license key. `kind`
        selects the shape: "individual" (default) locks to one email and one
        device id, redeemed once. "institution" locks to a verified-email
        domain (domainLock) with self-service seat management delegated to
        adminEmails — see the /v1/institutions/licenses/{licenseId}/seats
        routes below; Semper staff never touch individual seats on an
        institution key. The plaintext key is returned once, either way.
        "campus" is accepted as the pre-rename spelling of "institution".
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
        '422': { description: "kind=institution missing domainLock/adminEmails, or kind=individual missing emailLock/deviceIdLock" }
  /v1/admin/licenses/{licenseId}:
    patch:
      operationId: adminUpdateLicense
      description: >
        Device-attested, Semper-staff only. Change an existing license's terms
        in place — this is the renewal path. Extending expiresAt re-entitles
        every current holder (the individual redeemer, or every non-revoked
        institution seat) without issuing a new key or asking anyone to
        re-activate; the new terms are pushed to their user documents before
        the response returns. Terms only: kind, the email/device/domain locks
        and the key itself are fixed at mint. Send only the fields that change.
      parameters:
        - name: licenseId
          in: path
          required: true
          type: string
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
        '404': { description: License not found }
        '422': { description: "No field set, or a value out of range" }
        '429': { description: Rate limited }
  /v1/admin/licenses/{licenseId}/revoke:
    post:
      operationId: adminRevokeLicense
      description: >
        Device-attested, Semper-staff only. Revoke a key. Individual:
        the redeemer drops to demo. Institution: every seat on the license
        drops to demo and seatsUsed resets to 0 — this is the whole-key
        revoke; for offboarding a single institution member without affecting
        anyone else's seat, institution IT should use
        DELETE /v1/institutions/licenses/{licenseId}/seats/{uid} instead.
      parameters:
        - name: licenseId
          in: path
          required: true
          type: string
      responses:
        '200': { description: OK }
        '401': { description: Missing or invalid device attestation }
        '404': { description: License not found }
  # ---------------------------------------------------------------------------
  # Pre-rename aliases of the three routes above. ESPv2 is the front door and
  # rejects any path absent from this spec, so the aliases FastAPI serves are
  # unreachable in production unless they are declared here too — leaving them
  # out would make the deprecation window exist only against a directly-hit
  # Cloud Run URL, which nothing can reach. Same backend, same auth, same
  # handler. Delete this block and the FastAPI aliases together.
...